# PERF: pyemd to POT for EMD computation in `wmdistance`