--- a/docs/notebooks/WMD_tutorial.ipynb
+++ b/docs/notebooks/WMD_tutorial.ipynb
@@ -30,7 +30,7 @@
     "\n",
     "## Running this notebook\n",
     "\n",
-    "You can download this [iPython Notebook](http://ipython.org/notebook.html), and run it on your own computer, provided you have installed Gensim, PyEMD, NLTK, and downloaded the necessary data.\n",
+    "You can download this [iPython Notebook](http://ipython.org/notebook.html), and run it on your own computer, provided you have installed Gensim, POT, NLTK, and downloaded the necessary data.\n",
     "\n",
     "The notebook was run on an Ubuntu machine with an Intel core i7-4770 CPU 3.40GHz (8 cores) and 32 GB memory. Running the entire notebook on this machine takes about 3 minutes.\n",
     "\n",
@@ -524,8 +524,7 @@
    "source": [
     "## References\n",
     "\n",
-    "1. Ofir Pele and Michael Werman, *A linear time histogram metric for improved SIFT matching*, 2008.\n",
-    "* Ofir Pele and Michael Werman, *Fast and robust earth mover's distances*, 2009.\n",
+    "1. * Rémi Flamary et al. *POT: Python Optimal Transport*, 2021.\n",
     "* Matt Kusner et al. *From Embeddings To Document Distances*, 2015.\n",
     "* Thomas Mikolov et al. *Efficient Estimation of Word Representations in Vector Space*, 2013."
    ]