--- a/docs/notebooks/soft_cosine_tutorial.ipynb
+++ b/docs/notebooks/soft_cosine_tutorial.ipynb
@@ -30,7 +30,7 @@
     ">\n",
     "\n",
     "## Running this notebook\n",
-    "You can download this [Jupyter notebook](http://jupyter.org/), and run it on your own computer, provided you have installed the `gensim`, `jupyter`, `sklearn`, `pyemd`, and `wmd` Python packages.\n",
+    "You can download this [Jupyter notebook](http://jupyter.org/), and run it on your own computer, provided you have installed the `gensim`, `jupyter`, `sklearn`, `POT`, and `wmd` Python packages.\n",
     "\n",
     "The notebook was run on an Ubuntu machine with an Intel core i7-6700HQ CPU 3.10GHz (4 cores) and 16 GB memory. Assuming all resources required by the notebook have already been downloaded, running the entire notebook on this machine takes about 30 minutes."
    ]
@@ -357,7 +357,7 @@
    "metadata": {},
    "outputs": [],
    "source": [
-    "!pip install pyemd"
+    "!pip install POT"
    ]
   },
   {
@@ -404,7 +404,7 @@
     "    return similarities\n",
     "\n",
     "def wmd_gensim(query, documents):\n",
-    "    # Compute Word Mover's Distance as implemented in PyEMD by William Mayner\n",
+    "    # Compute Word Mover's Distance as implemented in POT\n",
     "    # between the query and the documents.\n",
     "    index = WmdSimilarity(documents, w2v_model)\n",
     "    similarities = index[query]\n",
@@ -532,26 +532,26 @@
    "cell_type": "markdown",
    "metadata": {},
    "source": [
-     "Dataset | Strategy | MAP score | Elapsed time (sec)\n",
-     ":---|:---|:---|---:\n",
-     "2016-test|softcossim|78.52 ±11.18|6.00 ±0.79\n",
-     "2016-test|**Winner (UH-PRHLT-primary)**|76.70 ±0.00|\n",
-     "2016-test|cossim|76.45 ±10.40|0.64 ±0.08\n",
-     "2016-test|wmd-gensim|76.23 ±11.42|5.37 ±0.64\n",
-     "2016-test|**Baseline 1 (IR)**|74.75 ±0.00|\n",
-     "2016-test|wmd-relax|71.05 ±11.06|1.11 ±0.09\n",
-     "2016-test|**Baseline 2 (random)**|46.98 ±0.00|\n",
-     "\n",
-     "\n",
-     "Dataset | Strategy | MAP score | Elapsed time (sec)\n",
-     ":---|:---|:---|---:\n",
-     "2017-test|**Winner (SimBow-primary)**|47.22 ±0.00|\n",
-     "2017-test|softcossim|45.88 ±16.22|7.08 ±1.49\n",
-     "2017-test|cossim|44.38 ±14.71|0.74 ±0.10\n",
-     "2017-test|wmd-gensim|44.06 ±15.92|6.20 ±0.87\n",
-     "2017-test|wmd-relax|43.52 ±16.30|1.30 ±0.18\n",
-     "2017-test|**Baseline 1 (IR)**|41.85 ±0.00|\n",
-     "2017-test|**Baseline 2 (random)**|29.81 ±0.00|"
+    "Dataset | Strategy | MAP score | Elapsed time (sec)\n",
+    ":---|:---|:---|---:\n",
+    "2016-test|softcossim|78.52 ±11.18|6.00 ±0.79\n",
+    "2016-test|**Winner (UH-PRHLT-primary)**|76.70 ±0.00|\n",
+    "2016-test|cossim|76.45 ±10.40|0.64 ±0.08\n",
+    "2016-test|wmd-gensim|76.23 ±11.42|5.37 ±0.64\n",
+    "2016-test|**Baseline 1 (IR)**|74.75 ±0.00|\n",
+    "2016-test|wmd-relax|71.05 ±11.06|1.11 ±0.09\n",
+    "2016-test|**Baseline 2 (random)**|46.98 ±0.00|\n",
+    "\n",
+    "\n",
+    "Dataset | Strategy | MAP score | Elapsed time (sec)\n",
+    ":---|:---|:---|---:\n",
+    "2017-test|**Winner (SimBow-primary)**|47.22 ±0.00|\n",
+    "2017-test|softcossim|45.88 ±16.22|7.08 ±1.49\n",
+    "2017-test|cossim|44.38 ±14.71|0.74 ±0.10\n",
+    "2017-test|wmd-gensim|44.06 ±15.92|6.20 ±0.87\n",
+    "2017-test|wmd-relax|43.52 ±16.30|1.30 ±0.18\n",
+    "2017-test|**Baseline 1 (IR)**|41.85 ±0.00|\n",
+    "2017-test|**Baseline 2 (random)**|29.81 ±0.00|"
    ]
   },
   {