--- a/docs/src/auto_examples/tutorials/run_fasttext.ipynb
+++ b/docs/src/auto_examples/tutorials/run_fasttext.ipynb
@@ -15,7 +15,7 @@
       "cell_type": "markdown",
       "metadata": {},
       "source": [
-        "\nFastText Model\n==============\n\nIntroduces Gensim's fastText model and demonstrates its use on the Lee Corpus.\n\n"
+        "\n# FastText Model\n\nIntroduces Gensim's fastText model and demonstrates its use on the Lee Corpus.\n"
       ]
     },
     {
@@ -40,14 +40,14 @@
       "cell_type": "markdown",
       "metadata": {},
       "source": [
-        "When to use fastText?\n---------------------\n\nThe main principle behind `fastText <https://github.com/facebookresearch/fastText>`_ is that the\nmorphological structure of a word carries important information about the meaning of the word.\nSuch structure is not taken into account by traditional word embeddings like Word2Vec, which\ntrain a unique word embedding for every individual word.\nThis is especially significant for morphologically rich languages (German, Turkish) in which a\nsingle word can have a large number of morphological forms, each of which might occur rarely,\nthus making it hard to train good word embeddings.\n\n\nfastText attempts to solve this by treating each word as the aggregation of its subwords.\nFor the sake of simplicity and language-independence, subwords are taken to be the character ngrams\nof the word. The vector for a word is simply taken to be the sum of all vectors of its component char-ngrams.\n\n\nAccording to a detailed comparison of Word2Vec and fastText in\n`this notebook <https://github.com/RaRe-Technologies/gensim/blob/develop/docs/notebooks/Word2Vec_FastText_Comparison.ipynb>`__,\nfastText does significantly better on syntactic tasks as compared to the original Word2Vec,\nespecially when the size of the training corpus is small. Word2Vec slightly outperforms fastText\non semantic tasks though. The differences grow smaller as the size of the training corpus increases.\n\n\nfastText can obtain vectors even for out-of-vocabulary (OOV) words, by summing up vectors for its\ncomponent char-ngrams, provided at least one of the char-ngrams was present in the training data.\n\n\n"
+        "## When to use fastText?\n\nThe main principle behind `fastText <https://github.com/facebookresearch/fastText>`_ is that the\nmorphological structure of a word carries important information about the meaning of the word.\nSuch structure is not taken into account by traditional word embeddings like Word2Vec, which\ntrain a unique word embedding for every individual word.\nThis is especially significant for morphologically rich languages (German, Turkish) in which a\nsingle word can have a large number of morphological forms, each of which might occur rarely,\nthus making it hard to train good word embeddings.\n\n\nfastText attempts to solve this by treating each word as the aggregation of its subwords.\nFor the sake of simplicity and language-independence, subwords are taken to be the character ngrams\nof the word. The vector for a word is simply taken to be the sum of all vectors of its component char-ngrams.\n\n\nAccording to a detailed comparison of Word2Vec and fastText in\n`this notebook <https://github.com/RaRe-Technologies/gensim/blob/develop/docs/notebooks/Word2Vec_FastText_Comparison.ipynb>`__,\nfastText does significantly better on syntactic tasks as compared to the original Word2Vec,\nespecially when the size of the training corpus is small. Word2Vec slightly outperforms fastText\non semantic tasks though. The differences grow smaller as the size of the training corpus increases.\n\n\nfastText can obtain vectors even for out-of-vocabulary (OOV) words, by summing up vectors for its\ncomponent char-ngrams, provided at least one of the char-ngrams was present in the training data.\n\n\n"
       ]
     },
     {
       "cell_type": "markdown",
       "metadata": {},
       "source": [
-        "Training models\n---------------\n\n\n"
+        "## Training models\n\n\n"
       ]
     },
     {
@@ -72,7 +72,7 @@
       "cell_type": "markdown",
       "metadata": {},
       "source": [
-        "Training hyperparameters\n^^^^^^^^^^^^^^^^^^^^^^^^\n\n\n"
+        "### Training hyperparameters\n\n\n"
       ]
     },
     {
@@ -93,7 +93,7 @@
       "cell_type": "markdown",
       "metadata": {},
       "source": [
-        "Saving/loading models\n---------------------\n\n\n"
+        "## Saving/loading models\n\n\n"
       ]
     },
     {
@@ -125,7 +125,7 @@
       "cell_type": "markdown",
       "metadata": {},
       "source": [
-        "Word vector lookup\n------------------\n\n\nAll information necessary for looking up fastText words (incl. OOV words) is\ncontained in its ``model.wv`` attribute.\n\nIf you don't need to continue training your model, you can export & save this `.wv`\nattribute and discard `model`, to save space and RAM.\n\n\n"
+        "## Word vector lookup\n\n\nAll information necessary for looking up fastText words (incl. OOV words) is\ncontained in its ``model.wv`` attribute.\n\nIf you don't need to continue training your model, you can export & save this `.wv`\nattribute and discard `model`, to save space and RAM.\n\n\n"
       ]
     },
     {
@@ -176,7 +176,7 @@
       "cell_type": "markdown",
       "metadata": {},
       "source": [
-        "Similarity operations\n---------------------\n\n\n"
+        "## Similarity operations\n\n\n"
       ]
     },
     {
@@ -230,7 +230,7 @@
       "cell_type": "markdown",
       "metadata": {},
       "source": [
-        "Other similarity operations\n^^^^^^^^^^^^^^^^^^^^^^^^^^^\n\nThe example training corpus is a toy corpus, results are not expected to be good, for proof-of-concept only\n\n"
+        "### Other similarity operations\n\nThe example training corpus is a toy corpus, results are not expected to be good, for proof-of-concept only\n\n"
       ]
     },
     {
@@ -292,7 +292,7 @@
       "cell_type": "markdown",
       "metadata": {},
       "source": [
-        "Word Movers distance\n^^^^^^^^^^^^^^^^^^^^\n\nYou'll need the optional ``pyemd`` library for this section, ``pip install pyemd``.\n\nLet's start with two sentences:\n\n"
+        "### Word Movers distance\n\nYou'll need the optional ``POT`` library for this section, ``pip install POT``.\n\nLet's start with two sentences:\n\n"
       ]
     },
     {
@@ -377,7 +377,7 @@
       "name": "python",
       "nbconvert_exporter": "python",
       "pygments_lexer": "ipython3",
-      "version": "3.6.5"
+      "version": "3.8.10"
     }
   },
   "nbformat": 4,