--- a/gensim/test/test_fasttext.py
+++ b/gensim/test/test_fasttext.py
@@ -30,10 +30,10 @@ from gensim.models.fasttext_inner import compute_ngrams, compute_ngrams_bytes, f
 import gensim.models.fasttext
 
 try:
-    from pyemd import emd  # noqa:F401
-    PYEMD_EXT = True
+    from ot import emd2  # noqa:F401
+    POT_EXT = True
 except (ImportError, ValueError):
-    PYEMD_EXT = False
+    POT_EXT = False
 
 logger = logging.getLogger(__name__)
 
@@ -394,7 +394,7 @@ class TestFastTextModel(unittest.TestCase):
         self.assertFalse('nights' in self.test_model.wv.key_to_index)
         self.assertTrue('nights' in self.test_model.wv)
 
-    @unittest.skipIf(PYEMD_EXT is False, "pyemd not installed")
+    @unittest.skipIf(POT_EXT is False, "POT not installed")
     def test_wm_distance(self):
         doc = ['night', 'payment']
         oov_doc = ['nights', 'forests', 'payments']