--- a/gensim/test/test_word2vec.py
+++ b/gensim/test/test_word2vec.py
@@ -21,10 +21,10 @@ import numpy as np
 from testfixtures import log_capture
 
 try:
-    from pyemd import emd  # noqa:F401
-    PYEMD_EXT = True
+    from ot import emd2  # noqa:F401
+    POT_EXT = True
 except (ImportError, ValueError):
-    PYEMD_EXT = False
+    POT_EXT = False
 
 from gensim import utils
 from gensim.models import word2vec, keyedvectors
@@ -1084,7 +1084,7 @@ class TestWord2VecModel(unittest.TestCase):
 
 class TestWMD(unittest.TestCase):
 
-    @unittest.skipIf(PYEMD_EXT is False, "pyemd not installed")
+    @unittest.skipIf(POT_EXT is False, "POT not installed")
     def test_nonzero(self):
         '''Test basic functionality with a test sentence.'''
 
@@ -1096,7 +1096,7 @@ class TestWMD(unittest.TestCase):
         # Check that distance is non-zero.
         self.assertFalse(distance == 0.0)
 
-    @unittest.skipIf(PYEMD_EXT is False, "pyemd not installed")
+    @unittest.skipIf(POT_EXT is False, "POT not installed")
     def test_symmetry(self):
         '''Check that distance is symmetric.'''
 
@@ -1107,7 +1107,7 @@ class TestWMD(unittest.TestCase):
         distance2 = model.wv.wmdistance(sentence2, sentence1)
         self.assertTrue(np.allclose(distance1, distance2))
 
-    @unittest.skipIf(PYEMD_EXT is False, "pyemd not installed")
+    @unittest.skipIf(POT_EXT is False, "POT not installed")
     def test_identical_sentences(self):
         '''Check that the distance from a sentence to itself is zero.'''