… docs/src/auto_examples/core/index.rst
… docs/src/auto_examples/howtos/index.rst
… docs/src/auto_examples/index.rst
… docs/src/auto_examples/other/index.rst
… docs/src/auto_examples/tutorials/images/sphx_glr_run_fasttext_001.png
… docs/src/auto_examples/tutorials/images/sphx_glr_run_wmd_001.png
… docs/src/auto_examples/tutorials/images/thumb/sphx_glr_run_fasttext_thumb.png
… docs/src/auto_examples/tutorials/images/thumb/sphx_glr_run_wmd_thumb.png
… docs/src/auto_examples/tutorials/index.rst
… docs/src/auto_examples/tutorials/run_fasttext.py.md5
… docs/src/auto_examples/tutorials/run_fasttext.rst
… docs/src/auto_examples/tutorials/run_fasttext_codeobj.pickle
… docs/src/auto_examples/tutorials/run_wmd.py.md5
… docs/src/auto_examples/tutorials/run_wmd.rst
… docs/src/auto_examples/tutorials/run_wmd_codeobj.pickle
… docs/src/auto_examples/tutorials/sg_execution_times.rst
… gensim/test/test_similarities.py