fixes #3312 

Why this change?

- Speed:
```
import gensim.downloader as api
corpus = api.load('text8')
model = api.load("glove-wiki-gigaword-50")
corpus = [d for i, d in enumerate(corpus) if i < 2]
```
Timings for the following call (took a subset because I'm on an old laptop right now):
```
model.wmdistance(corpus[0][:1000], corpus[1][:1000])
```
| pyemd (before)        | POT (after)           | 
| -------------------------- |:-----------------------:| 
| 30s                          | 84ms                   | 


- Provides other algorithms that could be used to introduce variants of the WMD (see discussion in #3312) 
- [Actively maintained](https://github.com/PythonOT/POT/releases): last release in December 2021 vs January 2018 for `pyemd`
- Has [good documentation](https://pythonot.github.io/)
- Same [MIT license](https://github.com/PythonOT/POT/blob/master/LICENSE)