@piskvorky 2022-04-16
Can you add some benchmarks into the PR description?

Something to point people to, a TL;DR for the release notes.