@sidravi1 2022-10-20
Thanks @TLouf! 

One other thing we may wish to do - 
there is a function provided by POT to calculate [distances](https://pythonot.github.io/all.html?highlight=dist#ot.dist). It falls back on [scipy's cdist](https://pythonot.github.io/_modules/ot/utils.html#dist) if metric is not euclidean.

Do we want to switch to that?