@gojomo 2022-10-21
As it, `ot.dist` replaces `cdist` on that line? I'd generally prefer to avoid extra changes unless/until there's a tangible benefit, in perf, functionality, or readability. 

(Perhaps, making distance-measure configurable improves performance in some cases. If we *know* that to be true via investigation, then it could be a good change... but even then, better as a focused PR with its own rationale, rather than mixed with this sort of drop-in, functionality-maintaining performance boost.)