@TLouf 2022-10-23
@mpenkov just went along CI's recommendation, however it added a ton of files, some of which I should not `git add` I think, like all the `index.rst` and the `.png`s. Let me know if that's correct, and if I should exclude more, and I'll revert the commit to only include what's necessary.